requirements.txt
absl-py==2.2.1
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
addict==2.4.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.1
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==3.0.0
astunparse==1.6.3
attrs==22.1.0
Babel==2.11.0
backports.tarfile==1.2.0
basicsr==1.3.4.9
beautifulsoup4==4.13.4
black==22.3.0
bleach==5.0.1
brotlipy==0.7.0
build==1.2.2.post1
builtin-interfaces==1.2.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
composition-interfaces==1.2.1
contourpy==1.3.0
cryptography==45.0.5
cv-bridge==3.2.1
cycler==0.12.1
debugpy==1.6.3
decorator==5.2.1
defusedxml==0.7.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
dill==0.3.9
DISTS-pytorch==0.1
docutils==0.21.2
domain-coordinator==0.10.0
einops==0.8.1
entrypoints==0.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.0.0
execnet==2.1.1
executing==2.2.0
expecttest==0.1.4
fastjsonschema==2.16.2
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.3.0
future==0.18.2
geometry-msgs==4.9.0
grpcio==1.71.0
-e git+https://github.com/XingtongGe/gsplat.git@bcca3ecae966a052e3bf8dd1ff9910cf7b8f851d#egg=gsplat
huggingface-hub==0.29.3
hypothesis==6.56.4
id==1.5.0
idna==3.10
image-geometry==3.2.1
imageio==2.37.0
importlib-metadata==5.0.0
importlib_resources==6.5.2
iniconfig==2.1.0
intel-cmplr-lib-ur==2025.2.0
intel-openmp==2025.2.0
interactive-markers==2.3.2
ipykernel==6.17.1
ipython==8.18.1
ipython-genutils==0.2.0
isort==5.10.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaxtyping==0.2.36
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
keyring==25.6.0
kiwisolver==1.4.7
laser-geometry==2.4.0
launch==1.0.9
launch-ros==0.19.9
launch-testing==1.0.9
launch-testing-ros==0.19.9
launch-xml==1.0.9
launch-yaml==1.0.9
lazy-object-proxy==1.10.0
lazy_loader==0.4
lifecycle-msgs==1.2.1
lmdb==1.6.2
logging-demo==0.20.5
lpips==0.1.4
map-msgs==2.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
message-filters==4.3.7
mistune==2.0.4
mkl==2025.2.0
mkl-service==2.4.0
mkl_fft==2.0.0
more-itertools==10.6.0
mpmath==1.2.1
mypy-extensions==1.0.0
nav-msgs==4.9.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
nh3==0.2.21
ninja==1.11.1.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.11.0.86
osrf-pycommon==2.1.6
packaging==24.2
pandocfilters==1.5.0
parso==0.8.4
pathspec==0.12.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus-client==0.15.0
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pycosat==0.6.3
pycparser==2.22
Pygments==2.19.1
pylint==2.13.4
pyparsing==3.0.9
pyproject_hooks==1.2.0
pyrsistent==0.19.2
pytest==7.1.2
pytest-forked==1.6.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0
pyzmq==24.0.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rfc3986==2.0.0
rich==13.9.4
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
SecretStorage==3.3.3
Send2Trash==1.8.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
sros2==0.10.6
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.11.1
tb-nightly==2.20.0a20250329
tbb==2022.2.0
tcmlib==1.4.0
teleop-twist-keyboard==2.4.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.17.0
tf2-geometry-msgs==0.25.13
tf2-kdl==0.25.13
tf2-msgs==0.25.13
tf2-py==0.25.13
tf2-ros-py==0.25.13
tf2-sensor-msgs==0.25.13
tf2-tools==0.25.13
tifffile==2024.8.30
timm==1.0.15
tinycss2==1.2.1
tokenize_rt==6.1.0
tomli==2.0.1
topic-monitor==0.20.5
torch==1.13.0
torchprofile==0.0.4
torchtext==0.14.0
torchvision==0.14.0
tornado==6.2
tqdm==4.67.1
traitlets==5.5.0
trajectory-msgs==4.9.0
turtlesim==1.4.2
twine==6.1.0
typeguard==4.4.2
types-dataclasses==0.6.6
typing_extensions==4.13.0
umf==0.11.0
unique-identifier-msgs==2.2.1
urllib3==2.5.0
visualization-msgs==4.9.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
zipp==3.10.0
